# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tracks api",
        "description": "Управление сущностями проекта трекера произведений.",
        "contact": {
            "email": "xenx.efy@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/async",
            "description": "Main server"
        }
    ],
    "paths": {
...
        "/tags": {
            "get": {
                "tags": [
                    "Tag"
                ],
                "summary": "Получение тегов.",
                "operationId": "getTags",
                "responses": {
                    "200": {
                        "description": "Success response"
                    }
                }
            },
            "post": {
                "tags": [
                    "Tag"
                ],
                "summary": "Создает тег.",
                "operationId": "storeTag",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title"
                                ],
                                "properties": {
                                    "title": {
                                        "description": "Название тега",
                                        "type": "string",
                                        "example": "Smooth"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success response"
                    },
                    "201": {
                        "description": "Success creation"
                    }
                }
            }
        },
        "/tags/{id}": {
            "put": {
                "tags": [
                    "Tag"
                ],
                "summary": "Обновляет информацию о теге.",
                "operationId": "updateTag",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id обновляемого тега",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Формат входящих данных",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "Название тега",
                                        "type": "string",
                                        "example": "Adazio"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success response"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tag"
                ],
                "summary": "Удалить тег.",
                "operationId": "deleteTag",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id удаляемого тега.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success response"
                    }
                }
            }
        },
        "/users": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Получение пользователей.",
                "operationId": "getUsers",
                "responses": {
                    "200": {
                        "description": "Success response"
                    }
                }
            },
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Создает пользователя",
                "operationId": "storeUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "is_admin"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "ФИО пользователя",
                                        "type": "string",
                                        "example": "Лебовский Василий Пупкин"
                                    },
                                    "email": {
                                        "description": "Email пользователя",
                                        "type": "string",
                                        "example": "lebovskiy@gmail.com"
                                    },
                                    "is_admin": {
                                        "description": "Признак администратора",
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success response"
                    },
                    "201": {
                        "description": "Success creation"
                    }
                }
            }
        },
        "/users/{id}": {
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Обновляет информацию о пользователе.",
                "operationId": "updateUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id обновляемого пользователя",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Формат входящих данных",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "ФИО пользователя.",
                                        "type": "string",
                                        "example": "Лебовский Василий Пупкин"
                                    },
                                    "email": {
                                        "description": "Email пользователя.",
                                        "type": "string",
                                        "example": "lebovskiy@gmail.com"
                                    },
                                    "is_admin": {
                                        "description": "Признак администратора.",
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success response"
                    }
                }
            },
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Удалить пользователя.",
                "operationId": "deleteUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id удаляемого пользователя.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success response"
                    }
                }
            }
        }
    },
...
    "security": [
        []
    ]
}